+++ packages/arcgis-rest-portal/test/items/get.test.ts
@@ -246,7 +246,7 @@ describe("get", () => {
       } as IAuthenticationManager;
       const item = await getItem("3ef", { authentication: fakeAuthManager });
       expect(item.thumbnailUrl).toBe(
-        "https://www.arcgis.com/sharing/rest/content/items/3ef/info/thumb.png?fake-token"
+        "https://www.arcgis.com/sharing/rest/content/items/3ef/info/thumb.png?token=fake-token"
       );
     });
 
@@ -257,6 +257,21 @@ describe("get", () => {
         portal: "https://www.arcgis.com/sharing/rest"
       } as IAuthenticationManager;
       const item = await getItem("3ef", { authentication: fakeAuthManager });
+      expect(item.thumbnailUrl).toBe(
+        "https://www.arcgis.com/sharing/rest/content/items/3ef/info/thumb.png"
+      );
+    });
+
+    test("should not decorate with an item thumbnail if there is an error getting a token", async () => {
+      fetchMock.once("*", MOCK_ITEM);
+      const fakeAuthManager = {
+        getToken: (_url: string) => Promise.reject(new Error("error")),
+        portal: "https://www.arcgis.com/sharing/rest",
+        federatedServers: []
+      } as IAuthenticationManager;
+      const item = await getItem("3ef", {
+        authentication: fakeAuthManager
+      });
       expect(item.thumbnailUrl).toBe(
         "https://www.arcgis.com/sharing/rest/content/items/3ef/info/thumb.png"
       );