+++ packages/arcgis-rest-auth/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @esri/arcgis-rest-auth Changelog
 
+## 4.11.0
+
+### Patch Changes
+
+- @esri/arcgis-rest-request@4.11.0
+
+## 4.10.3
+
+### Patch Changes
+
+- Updated dependencies [412188c]
+  - @esri/arcgis-rest-request@4.10.3
+
+## 4.10.2
+
+### Patch Changes
+
+- @esri/arcgis-rest-request@4.10.2
+
 ## 4.10.1
 
 ### Patch Changes