+++ packages/arcgis-rest-feature-service/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @esri/arcgis-rest-feature-service Changelog
 
+## 4.11.0
+
+### Minor Changes
+
+- 7923ffe: queryAttachments update
+
+## 4.10.3
+
+## 4.10.2
+
 ## 4.10.1
 
 ### Patch Changes